rename TabbedCarousel to ResourceCarousel
