+++ frontends/mit-open/src/page-components/ResourceCarousel/ResourceCarousel.test.tsx
@@ -1,11 +1,11 @@
 import React from "react"
-import TabbedCarousel from "./TabbedCarousel"
-import type { TabbedCarouselProps } from "./TabbedCarousel"
+import ResourceCarousel from "./ResourceCarousel"
+import type { ResourceCarouselProps } from "./ResourceCarousel"
 import { renderWithProviders, screen, user, waitFor } from "@/test-utils"
 import { factories, setMockResponse, makeRequest, urls } from "api/test-utils"
 import { act } from "@testing-library/react"
 
-describe("TabbedCarousel", () => {
+describe("ResourceCarousel", () => {
   const setupApis = () => {
     const resources = {
       search: factories.learningResources.resources({ count: 10 }),
@@ -23,7 +23,7 @@ describe("TabbedCarousel", () => {
   }
 
   it("returns results from the correct endpoint", async () => {
-    const config: TabbedCarouselProps["config"] = [
+    const config: ResourceCarouselProps["config"] = [
       {
         label: "Resources",
         data: {
@@ -42,7 +42,9 @@ describe("TabbedCarousel", () => {
 
     setMockResponse.get(urls.userMe.get(), {})
     const { list, search } = setupApis()
-    renderWithProviders(<TabbedCarousel title="My Carousel" config={config} />)
+    renderWithProviders(
+      <ResourceCarousel title="My Carousel" config={config} />,
+    )
     const tabs = screen.getAllByRole("tab")
     expect(tabs).toHaveLength(2)
     expect(tabs[0]).toHaveTextContent("Resources")
@@ -58,29 +60,7 @@ describe("TabbedCarousel", () => {
   }, 10000)
 
   it("calls API with expected parameters", async () => {
-    const config: TabbedCarouselProps["config"] = [
-      {
-        label: "Resources",
-        data: {
-          type: "resources",
-          params: { resource_type: ["course", "program"], professional: true },
-        },
-      },
-    ]
-    setMockResponse.get(urls.userMe.get(), {})
-    setupApis()
-    renderWithProviders(<TabbedCarousel title="My Carousel" config={config} />)
-    await waitFor(() => {
-      expect(makeRequest.mock.calls.length > 0).toBe(true)
-    })
-    const [_method, url] = makeRequest.mock.calls[0]
-    const urlParams = new URLSearchParams(url.split("?")[1])
-    expect(urlParams.getAll("resource_type")).toEqual(["course", "program"])
-    expect(urlParams.get("professional")).toEqual("true")
-  })
-
-  it("Shows the correct title", () => {
-    const config: TabbedCarouselProps["config"] = [
+    const config: ResourceCarouselProps["config"] = [
       {
         label: "Resources",
         data: {
@@ -92,7 +72,31 @@ describe("TabbedCarousel", () => {
     setMockResponse.get(urls.userMe.get(), {})
     setupApis()
     renderWithProviders(
-      <TabbedCarousel title="My Favorite Carousel" config={config} />,
+      <ResourceCarousel title="My Carousel" config={config} />,
+    )
+    await waitFor(() => {
+      expect(makeRequest.mock.calls.length > 0).toBe(true)
+    })
+    const [_method, url] = makeRequest.mock.calls[0]
+    const urlParams = new URLSearchParams(url.split("?")[1])
+    expect(urlParams.getAll("resource_type")).toEqual(["course", "program"])
+    expect(urlParams.get("professional")).toEqual("true")
+  })
+
+  it("Shows the correct title", () => {
+    const config: ResourceCarouselProps["config"] = [
+      {
+        label: "Resources",
+        data: {
+          type: "resources",
+          params: { resource_type: ["course", "program"], professional: true },
+        },
+      },
+    ]
+    setMockResponse.get(urls.userMe.get(), {})
+    setupApis()
+    renderWithProviders(
+      <ResourceCarousel title="My Favorite Carousel" config={config} />,
     )
     expect(
       screen.getByRole("heading", { name: "My Favorite Carousel" }),
