+++ frontends/mit-open/src/page-components/ResourceCarousel/types.ts
